Setup: factor out a base objective for metric collection
